DataBase Scripts
Scripts that connects to the database and get and set data. Everything is explained both in code comments and in a txt file called "documentation". This also includes php files for connection to data base. VERY IMPORTANT! The code is writen to connect to a local database. This can be changed in two simple steps, but must be changed when the database server is ready.